fix sv global configuration
